adk-agent/pyproject.toml
[project]
name = "google-adk-project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[database]==0.3.0",
    "yfinance==0.2.56",
    "psutil==5.9.5",
    "litellm==1.66.3",
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.0",
    "deprecated>=1.2.18",
]
